V.1.7

+++ routes/web.php
@@ -3,6 +3,7 @@
 use App\Models\Listing;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\UserController;
 use App\Http\Controllers\ListingController;
 
 Route::get('/', function () {
@@ -41,16 +42,26 @@ destroy: delete teh listing
 
 Route::get('/', [ListingController::class, 'index']);
 
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
-Route::post('/listings', [ListingController::class, 'store']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
-Route::get('listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 /* Route::put('/listings/{listing}/logo', [ListingController::class, 'setLogoNull']);
  */
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
+
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');//show register form
+
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+
+Route::post('/users', [UserController::class, 'store']);
+
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
+
+Route::post('/users/authenticate', [UserController::class, 'authenticate']);
